[FRONTEND] New kernel launch API.
Previously compiler flags were passed in the same list of parameters as
kernel arguments, as in

  my_kernel[grid](arg1, arg2, ..., num_warps=42).

This is problematic for a variety of reasons, but perhaps principal
among them is the fact that we cannot add new compiler flags without
potentially breaking kernels that use the arg name that we now want to
claim.

In the new syntax, you write

  my_kernel.with_flags(num_warps=42)[grid](arg1, arg2, ...)

In addition to fixing the problem of adding new compiler flags, this
also plays nicer with Python type-checkers.  Previously, pylance would
incorrectly report that `num_warps` was not an argument to the kernel
function, whereas now `with_flags` explicitly declares all its
parameters in a manner friendly to the type checker.

GPC: new-kernel-launch-api